Added so that the delete function also checks to se if there is a product with the correct id before deleting it. Thus requireing a findbyid function that I also added to the routes function

+++ vite-project/serverRest/routes/productRoute.js
@@ -1,6 +1,10 @@
 const express = require("express");
 const router = express.Router();
-const { createProduct, deleteProduct } = require("../db/productService");
+const {
+  createProduct,
+  deleteProduct,
+  findProductById,
+} = require("../db/productService");
 
 // POST
 router.post("/", async (request, response) => {
@@ -17,10 +21,16 @@ router.delete("/:id", async (req, res) => {
       res.status(400).send("ID is required");
       return;
     }
+
+    const product = await findProductById(id);
+    if (!product) {
+      res.status(404).send("Product not found");
+      return;
+    }
     await deleteProduct(id);
-    res.status(204).send("ID deleted");
+    res.status(204).send();
   } catch (error) {
-    res.status(500).send("internal server error", error);
+    res.status(500).send(error.message);
   }
 });
